geo_server/pyproject.toml
[project]
name = "geo-server"
version = "0.1.0"
description = "GEO Server - 后端服务：Web 自动化与数据采集"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "playwright>=1.40.0",
    "psycopg2-binary>=2.9.9",
    "tldextract>=5.1.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0.1",
    "tabulate>=0.9.0",
    "jieba>=0.42.1",
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "requests>=2.31.0",
    "bcrypt>=4.0.0",
    "PyJWT>=2.8.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
include = [
    "api.py",
    "main.py", 
    "stats.py",
    "stats_full.py",
    "config.yaml",
    "core",
    "providers",
]

[tool.uv]
dev-dependencies = []

